feat: Limit SEP review comment characters to 2000
